requirements.txt
Django==5.0.1
djangorestframework==3.14.0
mysqlclient==2.2.6


